requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
motor>=3.6,<4
httpx>=0.27,<1
PyJWT[crypto]>=2.9,<3
python-dotenv>=1.0,<2

